refactor:重构上传文件工具类(李洪军)

+++ common/media/src/main/java/util/FastDFSUploadUtils.java
@@ -8,8 +8,6 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.Collection;
-import java.util.HashMap;
-import java.util.Map;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.Part;
@@ -20,6 +18,10 @@ import org.slf4j.LoggerFactory;
 import org.springframework.util.StringUtils;
 
 import com.lawu.eshop.utils.RandomUtil;
+import util.upload.FastDFSClient;
+import util.upload.FastDFSResult;
+import util.upload.FastDFSResultEnum;
+import util.upload.FileUploadConstant;
 
 /**
  * FastDFS上传工具
@@ -33,7 +35,6 @@ public class FastDFSUploadUtils {
 
 	public static FastDFSResult upload(HttpServletRequest request, UploadParam uparam) {
 		FastDFSResult result = new FastDFSResult();
-		Map<String, String> map = new HashMap<String, String>();
 		Collection<Part> parts = null;
 		try {
 			parts = request.getParts();
@@ -46,21 +47,21 @@ public class FastDFSUploadUtils {
 			InputStream in = null;
 			if (part.getContentType() != null) {
 				long fileSize = part.getSize();
-				if (fileSize > 0 && fileSize < 50 * 1048576) {
+				if (fileSize > 0 && fileSize < FileUploadConstant.VIEDO_MAX_SIZE) {
 					String fileName = part.getSubmittedFileName();
 					String extName = fileName.substring(fileName.lastIndexOf(".") + 1);
 					try {
 						in = part.getInputStream();
 						ClientParams param = uparam.getCparam();
-						if (uparam.getType().equals("img")) {// 图片文件
-							if (fileSize > 5 * 1048576) {
+						if (uparam.getType().equals(FileUploadConstant.IMG)) {// 图片文件
+							if (fileSize > FileUploadConstant.IMG_MAX_SIZE) {
 								result.setFenum(FastDFSResultEnum.FD_FILE_IMG_BIG);
 								return result;
 							}
 							String fileUrl = FastDFSClient.getInstance(param)
 									.uploadFile(FastDFSClient.getFileBuffer(in, fileSize), extName);
 							result.setFileUrl(fileUrl);
-						} else if (uparam.getType().equals("viedo")) {// 视频文件
+						} else if (uparam.getType().equals(FileUploadConstant.VIEDO)) {// 视频文件
 							if (StringUtils.isEmpty(uparam.getFfmpegUrl())) {
 								result.setFenum(FastDFSResultEnum.FD_FILE_ERROR);
 								return result;
@@ -69,7 +70,6 @@ public class FastDFSUploadUtils {
 							saveLocalFile(part, uparam, newName);
 							String fileUrl = FastDFSClient.getInstance(param)
 									.uploadFile(FastDFSClient.getFileBuffer(in, fileSize), extName);
-							map.put("videoUrl", fileUrl);
 							result.setFileUrl(fileUrl);
 							String tmpVideoUrl = uparam.getBaseImageDir() + File.separator + uparam.getDir()
 									+ File.separator + newName;
@@ -82,29 +82,42 @@ public class FastDFSUploadUtils {
 							}
 							result.setCutImgUrl(cutImgUrl);
 							File redeoFile = new File(tmpVideoUrl);
-							redeoFile.delete();
+							if (redeoFile.exists()) {
+								redeoFile.delete();
+							}
 							File cutImgFile = new File(cutUrl);
-							cutImgFile.delete();
+							if (cutImgFile.exists()) {
+								cutImgFile.delete();
+							}
 						} else {// 其他文件
 							String fileUrl = FastDFSClient.getInstance(param)
 									.uploadFile(FastDFSClient.getFileBuffer(in, fileSize), extName);
 							result.setFileUrl(fileUrl);
 						}
-						in.close();
 						result.setFenum(FastDFSResultEnum.FD_UPLOAD_SUCCESS);
 						return result;
 					} catch (Exception e) {
 						log.error("FastDFS上传文件异常" + e.getMessage());
 						try {
-							in.close();
+							if (null != in) {
+								in.close();
+							}
 						} catch (IOException e1) {
 							log.error("关闭流异常" + e.getMessage());
 						}
 						result.setFenum(FastDFSResultEnum.FD_FILE_ERROR);
 						return result;
+					} finally {
+						try {
+							if (null != in) {
+								in.close();
+							}
+						} catch (Exception e) {
+							log.error("关闭流异常" + e.getMessage());
+						}
 					}
 				} else {
-					log.error("上传文件大于50M");
+					log.error("上传文件大于" + FileUploadConstant.VIEDO_MAX_SIZE + "M");
 					result.setFenum(FastDFSResultEnum.UPLOAD_SIZE_BIGER);
 					return result;
 				}
@@ -125,14 +138,21 @@ public class FastDFSUploadUtils {
 		if (!file.exists()) {
 			file.mkdirs();
 		}
-		byte[] b = new byte[1024 * 1024];
 		FileOutputStream out = new FileOutputStream(new File(file, newName));
-		int index = 0;
-		InputStream fileIn = part.getInputStream();
-		while ((index = fileIn.read(b)) != -1) {
-			out.write(b, 0, index);
+		byte[] b = new byte[1024 * 1024];
+		try{
+			int index = 0;
+			InputStream fileIn = part.getInputStream();
+			while ((index = fileIn.read(b)) != -1) {
+				out.write(b, 0, index);
+			}
+		}catch (IOException e){
+			throw e;
+		}finally {
+			if(null != out){
+				out.close();
+			}
 		}
-		out.close();
 	}
 
 }
